Move services to service directory.
