+++ lib/blocs/listeningTabContentBloc.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:readr_app/helpers/apiResponse.dart';
 import 'package:readr_app/helpers/constants.dart';
-import 'package:readr_app/models/listeningTabContentService.dart';
+import 'package:readr_app/services/listeningTabContentService.dart';
 import 'package:readr_app/models/recordList.dart';
 
 class ListeningTabContentBloc {
