+++ lib/blocs/tabContentBloc.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:readr_app/helpers/apiResponse.dart';
 import 'package:readr_app/helpers/constants.dart';
-import 'package:readr_app/models/editorChoiceService.dart';
+import 'package:readr_app/services/editorChoiceService.dart';
 import 'package:readr_app/models/recordList.dart';
-import 'package:readr_app/models/recordService.dart';
+import 'package:readr_app/services/recordService.dart';
 
 class TabContentBloc {
   RecordList _records;
