+++ lib/services/newsMarqueeService.dart
